Add server-side puzzle fetch and API
